@@ -1,6 +1,6 @@
 [metadata]
 name = napistu-torch
-version = 0.1.0
+version = 0.1.1
 description = PyTorch-based toolkit for working with Napistu network graphs
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -13,8 +13,6 @@ classifiers =
     Operating System :: OS Independent
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
 
 [options]
@@ -26,7 +24,7 @@ install_requires =
     pandas>=1.3.0
     pydantic>=2.0.0
     scikit-learn>=1.0.0
-python_requires = >=3.9
+python_requires = >=3.11
 package_dir =
     = src
 
@@ -39,10 +37,13 @@ pyg =
     torch-sparse>=0.6.17
     torch-scatter>=2.1.1
 dev =
-    pytest>=7.0.0
     black>=22.0.0
+    pytest>=7.0.0
     ruff>=0.1.0
+    torch-geometric>=2.4.0
+    torch-sparse>=0.6.17
+    torch-scatter>=2.1.1
 
 [tool:pytest]
-testpaths = tests
-addopts = --disable-warnings
+filterwarnings =
+    ignore::DeprecationWarning
